Recocognize player team, ball interpolation, ball assignment
Added detection for a players team based on his t-shirt color using KMeans. Now in the video output each player has a distinct visual cue for the team he's in.

Added the posibility to interpolate the balls position throughout the frames, though it's not performing as good as I had hoped.

Added ball assignment, now a player can be assigned the ball possession if either of his feet are close enough to the ball's bounding box.